src/main/java/com/vkumar/userservice/controller/UserController.java
package com.vkumar.userservice.controller;

import com.vkumar.userservice.models.UserData;
import com.vkumar.userservice.services.UserService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/users")
public class UserController {

    @Autowired
    private UserService userService;

    @PostMapping
    public ResponseEntity<UserData> createUser(@RequestBody UserData data) {
        UserData result = userService.addUser(data);
        return ResponseEntity.status(HttpStatus.CREATED).body(result);
    }

    @GetMapping
    public ResponseEntity<List<UserData>> getAllUsers() {
        return ResponseEntity.ok(userService.getAlluser());
    }

    @GetMapping("/{userId}")
    public ResponseEntity<UserData> getUser(@PathVariable("userId") String userId) {
        return ResponseEntity.ok(userService.getUser(userId));
    }
}
